pyproject.toml
[project]
name = "ccl25-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
]
